requirements.txt
google-cloud-optimization
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
protobuf
pytz
Flask
python-dotenv
json
